requirements.txt
PyPDF2
cohere
openpyxl
python-dotenv
python-docx
streamlit







